+++ sdk-src/src/main/java/com/hpe/adm/nga/sdk/network/OctaneUrl.java
@@ -90,6 +90,9 @@ public final class OctaneUrl {
 
         //Append paths
         if(getPaths().size()>0){
+            if (!url.endsWith("/")) {
+                url += "/";
+            }
             url += getPaths().stream().collect(Collectors.joining(PATH_SEPARATOR));
         }
