+++ sdk-src/src/test/java/com/hpe/adm/nga/sdk/model/TestQuery.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.hpe.adm.nga.sdk.unit_tests.model;
+package com.hpe.adm.nga.sdk.model;
 
 import static org.junit.Assert.*;
