--- sdk-src/src/test/java/com/hpe/adm/nga/sdk/unit_tests/QuickExample.java
@@ -1,26 +1,0 @@
-/*
- *
- *    Copyright 2017 Hewlett-Packard Development Company, L.P.
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-package com.hpe.adm.nga.sdk.unit_tests;
-
-import com.hpe.adm.nga.sdk.Octane;
-import com.hpe.adm.nga.sdk.authentication.SimpleUserAuthentication;
-
-public class QuickExample {
-
-    public static void main (String [] args) throws Exception {
-        Octane octane = new Octane.Builder(new SimpleUserAuthentication("sa@nga", "Welcome1")).build();
-    }
-}
